+++ apps/backend/modules/system/index.ts
@@ -1,7 +1,1 @@
-export {
-  INCOMING_SUPPLIER_RESPONSIBLE_DEPARTMENT_ID_SETTING_KEY,
-  InspectionRequestResponsibilityDepartmentSettingService,
-  PROCESS_OUTSOURCING_RESPONSIBLE_DEPARTMENT_ID_SETTING_KEY,
-  ProcessOutsourcingResponsibleDepartmentSettingService,
-} from './inspection-request-outsourcing-responsibility-setting.service';
 export { SystemService } from './system.service';